\name{humanFormat}
\alias{humanFormat}
\docType{package}
\title{
Format numbers into human readable strings
}
\description{
When working with durations, bytes, and other items with odd bases, it's
often useful to format the numbers into human readable units.
}
\details{
  \itemize{
  \item \code{\link{formatDuration}}: format durations
   \item \code{\link{formatBytes}}: format byte sizes
  }
\tabular{ll}{
Package: \tab humanFormat\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-11-22\cr
License: \tab MIT\cr
}
}
\author{
Dustin Sallings

Maintainer: Dustin Sallings <dustin@spy.net>
}
\keyword{ package
format
label
}
