\name{humanFormat}
\alias{humanFormat}
\docType{package}
\title{
Format numbers into human readable strings
}
\description{
When working with durations, bytes, and other items with odd bases, it's
often useful to format the numbers into human readable units.
}
\details{
  \itemize{
  \item \code{\link{formatDuration}}: format durations
   \item \code{\link{formatBytes}}: format byte sizes
  }
\tabular{ll}{
Package: \tab humanFormat\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2016-05-13\cr
License: \tab MIT\cr
BugReports: \tab https://github.com/dustin/humanFormat/issues\cr
}
}
\author{
Dustin Sallings

Maintainer: Dustin Sallings <dustin@spy.net>
}
\keyword{format}
